+++ src/urllib3/_async/connectionpool.py
@@ -42,7 +42,13 @@ from ..util.ssl_ import (
     BaseSSLError,
 )
 from ..util.timeout import Timeout
-from ..util.url import get_host, Url, _normalize_host as normalize_host
+from ..util.url import (
+    get_host,
+    parse_url,
+    Url,
+    _normalize_host as normalize_host,
+    _encode_target,
+)
 from ..util.queue import LifoQueue
 
 try:
@@ -581,6 +587,12 @@ class HTTPConnectionPool(ConnectionPool, RequestMethods):
 
         if not isinstance(retries, Retry):
             retries = Retry.from_int(retries, default=self.retries, redirect=False)
+
+        # Ensure that the URL we're connecting to is properly encoded
+        if url.startswith("/"):
+            url = six.ensure_str(_encode_target(url))
+        else:
+            url = six.ensure_str(parse_url(url).url)
 
         conn = None